+++ src/tests.rs
@@ -14,10 +14,10 @@ use crate::config;
 use crate::instruction;
 use crate::packet::{payload, sync};
 use crate::tracer;
-use crate::types::{branch, stack, trap};
+use crate::types::{branch, stack, trap, Context};
 
 use instruction::{Kind, COMPRESSED, UNCOMPRESSED};
-use tracer::item::{Context, Item};
+use tracer::item::Item;
 
 trace_test!(
     full_branch_map,