+++ src/tracer/item.rs
@@ -3,8 +3,7 @@
 //! Tracing item
 
 use crate::instruction::{self, info, Instruction};
-use crate::packet::sync;
-use crate::types::{trap, Privilege};
+use crate::types::{trap, Context};
 
 /// Tracing item
 ///
@@ -96,29 +95,5 @@ impl<I: info::Info> From<trap::Info> for Kind<I> {
 impl<I: info::Info> From<Context> for Kind<I> {
     fn from(context: Context) -> Self {
         Self::Context(context)
-    }
-}
-
-/// Execution context
-#[derive(Copy, Clone, Debug, Default, Eq, PartialEq, Hash)]
-pub struct Context {
-    /// The privilege level under which code is executed
-    pub privilege: Privilege,
-    /// The context of the execution
-    pub context: u64,
-}
-
-impl From<&sync::Context> for Context {
-    fn from(ctx: &sync::Context) -> Self {
-        Self {
-            privilege: ctx.privilege,
-            context: ctx.context.unwrap_or_default(),
-        }
-    }
-}
-
-impl From<sync::Context> for Context {
-    fn from(ctx: sync::Context) -> Self {
-        (&ctx).into()
     }
 }