+++ CHANGELOG.md
@@ -32,6 +32,7 @@ is based on https://keepachangelog.com/en/1.1.0/.
 - The `tracer::stack` module was moved to `types`.
 - `Iterator::size_hint` is now implemented explicitly (and not as a provided fn)
   for `tracer::Tracer`.
+- `tracer::item::Context` was moved to `types`.
 
 ### Added